requirements.txt
streamlit
pymongo
PyPDF2
python-docx
openai
google-generativeai
llama-index
werkzeug
numpy
pandas
plotly
scikit-learn
networkx
community
umap-learn
seaborn
matplotlib
scipy
Pillow
python-dotenv
zoomus
asyncio
google-auth-oauthlib
google-auth
transformers
textstat
spacy
streamlit_option_menu
beautifulsoup4
youtube-transcript-api
requests
xml==0.0.1
networkx==3.1
bokeh==3.2.1
scikit-learn==1.2.2
langchain==0.0.208